+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+# Enhancements
+
+* Adds the `SpeculativeEnabled` field to the `StackCreateOptions` and `StackUpdateOptions` structs by @arunatibm [1279](https://github.com/hashicorp/go-tfe/pull/1279)
+
 # v1.100.0
 
 ## Enhancements